+++ database/migrations/2016_11_23_125510_create_lt_Fight_table.php
@@ -14,37 +14,41 @@ class CreateLtFightTable extends Migration
      */
     public function up()
     {
-        Schema::create('fight', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('short_id')->unsigned()->nullable();
-            $table->integer('fighters_group_id')->unsigned()->index();
-            $table->foreign('fighters_group_id')
-                ->references('id')
-                ->on('fighters_groups')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->integer('c1')->nullable()->unsigned()->index();
-            $table->integer('c2')->nullable()->unsigned()->index();
-            $table->char('point1_c1')->nullable();
-            $table->char('point2_c1')->nullable();
-            $table->char('point1_c2')->nullable();
-            $table->char('point2_c2')->nullable();
-            $table->integer('winner_id')->unsigned()->nullable();
+        if (!Schema::hasTable('fight')) {
+            Schema::create('fight', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('short_id')->unsigned()->nullable();
+                $table->integer('fighters_group_id')->unsigned()->index();
+                $table->foreign('fighters_group_id')
+                    ->references('id')
+                    ->on('fighters_groups')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->integer('c1')->nullable()->unsigned()->index();
+                $table->integer('c2')->nullable()->unsigned()->index();
+                $table->char('point1_c1')->nullable();
+                $table->char('point2_c1')->nullable();
+                $table->char('point1_c2')->nullable();
+                $table->char('point2_c2')->nullable();
+                $table->integer('winner_id')->unsigned()->nullable();
 
-            $table->boolean('hansoku1_c1')->nullable();
-            $table->boolean('hansoku2_c1')->nullable();
-            $table->boolean('hansoku3_c1')->nullable();
-            $table->boolean('hansoku4_c1')->nullable();
-            $table->boolean('hansoku1_c2')->nullable();
-            $table->boolean('hansoku2_c2')->nullable();
-            $table->boolean('hansoku3_c2')->nullable();
-            $table->boolean('hansoku4_c2')->nullable();
+                $table->boolean('hansoku1_c1')->nullable();
+                $table->boolean('hansoku2_c1')->nullable();
+                $table->boolean('hansoku3_c1')->nullable();
+                $table->boolean('hansoku4_c1')->nullable();
+                $table->boolean('hansoku1_c2')->nullable();
+                $table->boolean('hansoku2_c2')->nullable();
+                $table->boolean('hansoku3_c2')->nullable();
+                $table->boolean('hansoku4_c2')->nullable();
 
-            $table->tinyInteger('area')->default(1);
-            $table->tinyInteger('order')->default(1);
-            $table->timestamps();
-            $table->engine = 'InnoDB';
-        });
+                $table->tinyInteger('area')->default(1);
+                $table->tinyInteger('order')->default(1);
+                $table->timestamps();
+                $table->engine = 'InnoDB';
+            });
+        }
+
+
     }
 
     /**
