+++ database/migrations/2017_03_15_000438_create_lt_Competitor_Team_table.php
@@ -14,27 +14,30 @@ class CreateLtCompetitorTeamTable extends Migration
      */
     public function up()
     {
-        Schema::create('competitor_team', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('team_id')->unsigned()->nullable()->index();
-            $table->integer('competitor_id')->unsigned()->index();
-            $table->integer('order')->unsigned()->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('fighters_group_team')) {
+            Schema::create('fighters_group_team', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('team_id')->unsigned()->nullable()->index();
+                $table->integer('competitor_id')->unsigned()->index();
+                $table->integer('order')->unsigned()->nullable();
+                $table->timestamps();
 
-            $table->foreign('team_id')
-                ->references('id')
-                ->on('team')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->foreign('team_id')
+                    ->references('id')
+                    ->on('team')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-            $table->foreign('competitor_id')
-                ->references('id')
-                ->on('competitor')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->foreign('competitor_id')
+                    ->references('id')
+                    ->on('competitor')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-            $table->unique(['team_id', 'competitor_id']);
-        });
+                $table->unique(['team_id', 'competitor_id']);
+            });
+        }
+
     }
 
     /**
